Remove header scrollbar so that it doesn't hides annotations (#11)
* Remove browser scrollbar and use JS to handle scrolling
* Refactor scrolling logic
* Fix mouse events not being fired
* Avoids viewer breaking when scroll offset is 0
